+++ databallpy/load_data/metadata.py
@@ -26,6 +26,7 @@ class Metadata:
     away_formation: str
 
     country: str
+    periods_changed_playing_direction: list = None
 
     def __post_init__(self):
         # match id
@@ -58,15 +59,15 @@ class Metadata:
                 f"periods_frames should be a pandas dataframe, not a \
                     {type(self.periods_frames)}"
             )
-        if "period" not in self.periods_frames.columns:
+        if "period_id" not in self.periods_frames.columns:
             raise ValueError("'period' should be one of the columns in period_frames")
         if any(
             [
-                x not in self.periods_frames["period"].value_counts().index
+                x not in self.periods_frames["period_id"].value_counts().index
                 for x in [1, 2, 3, 4, 5]
             ]
-        ) or not all(self.periods_frames["period"].value_counts() == 1):
-            res = self.periods_frames["period"]
+        ) or not all(self.periods_frames["period_id"].value_counts() == 1):
+            res = self.periods_frames["period_id"]
             raise ValueError(
                 f"'period' column in period_frames should contain only the values \
                     [1, 2, 3, 4, 5]. Now it's {res}"
@@ -190,6 +191,10 @@ class Metadata:
                 else pd.isnull(other.away_score),
                 self.away_formation == other.away_formation,
                 self.country == other.country,
+                self.periods_changed_playing_direction
+                == other.periods_changed_playing_direction
+                if self.periods_changed_playing_direction is not None
+                else other.periods_changed_playing_direction is None,
             ]
             return all(result)
         else:
@@ -212,4 +217,9 @@ class Metadata:
             away_score=self.away_score,
             away_formation=self.away_formation,
             country=self.country,
+            periods_changed_playing_direction=list(
+                self.periods_changed_playing_direction
+            )
+            if self.periods_changed_playing_direction is not None
+            else None,
         )