+++ tests/load_data/event_data/test_opta.py
@@ -19,6 +19,7 @@ from databallpy.load_data.event_data.opta import (
 from databallpy.load_data.event_data.pass_event import PassEvent
 from databallpy.load_data.event_data.shot_event import ShotEvent
 from databallpy.utils.utils import MISSING_INT
+from databallpy.warnings import DataBallPyWarning
 from tests.expected_outcomes import (
     DRIBBLE_EVENTS_OPTA,
     ED_OPTA,
@@ -32,6 +33,9 @@ class TestOpta(unittest.TestCase):
     def setUp(self):
         self.f7_loc = "tests/test_data/f7_test.xml"
         self.f7_loc_no_timestamps = "tests/test_data/f7_test_no_timestamps.xml"
+        self.f7_loc_no_timestamps_and_date = (
+            "tests/test_data/f7_test_no_timestamps_and_date.xml"
+        )
         self.f7_loc_multiple_matches = "tests/test_data/f7_test_multiple_matches.xml"
         self.f24_loc = "tests/test_data/f24_test.xml"
 
@@ -54,6 +58,14 @@ class TestOpta(unittest.TestCase):
             assert key in shot_events_opta.keys()
             assert event == shot_events_opta[key]
 
+    def test_load_event_data_pass(self):
+        event_data, metadata, dbp_events = load_opta_event_data(
+            self.f7_loc, "pass", pitch_dimensions=[10.0, 10.0]
+        )
+        assert metadata == MD_OPTA
+        pd.testing.assert_frame_equal(event_data, pd.DataFrame())
+        assert dbp_events == {}
+
     def test_load_metadata(self):
         metadata = _load_metadata(self.f7_loc, [10.0, 10.0])
         assert metadata == MD_OPTA
@@ -63,9 +75,25 @@ class TestOpta(unittest.TestCase):
         # the second match metadata is dropped
         assert metadata == MD_OPTA
 
-    def test_load_metadata_no_timestamps(self):
+    def test_load_metadata_no_timestamps_and_date(self):
         with self.assertRaises(ValueError):
-            _load_metadata(self.f7_loc_no_timestamps, [10.0, 10.0])
+            _load_metadata(self.f7_loc_no_timestamps_and_date, [10.0, 10.0])
+
+    def test_load_metadata_no_timestamps(self):
+        with self.assertWarns(DataBallPyWarning):
+            metadata = _load_metadata(self.f7_loc_no_timestamps, [10.0, 10.0])
+        expected_md = MD_OPTA.copy()
+        pf = expected_md.periods_frames
+        pf.loc[0, "start_datetime_ed"] = pd.to_datetime(
+            "20230122T111500+0000", utc=True
+        )
+        pf.loc[0, "end_datetime_ed"] = pd.to_datetime("20230122T120000+0000", utc=True)
+        pf.loc[1, "start_datetime_ed"] = pd.to_datetime(
+            "20230122T121500+0000", utc=True
+        )
+        pf.loc[1, "end_datetime_ed"] = pd.to_datetime("20230122T130000+0000", utc=True)
+
+        assert metadata == expected_md
 
     def test_get_player_info(self):
         player_data = [