+++ tests/load_data/tracking_data/utils/test_add_ball_data_to_dict.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from databallpy.load_data.tracking_data._add_ball_data_to_dict import (
+from databallpy.load_data.tracking_data.utils._add_ball_data_to_dict import (
     _add_ball_data_to_dict,
 )